requirements.txt
Django==1.4.3
boto==2.6.0
dj-database-url==0.2.1
django-appconf==0.5
django-compressor==1.2
django-debug-toolbar==0.9.4
django-storages==1.1.5
gunicorn==0.16.1
psycopg2==2.4.5
wsgiref==0.1.2
